Enable drag at top, or twist gesture, to change camera rotation
Modify GestureResponses.cs to work with the new FollowCamera script, and enable rotating the view relative to the target, with either a twist gesture on iOS, or a horizontal drag on the dummy camera at the top of the view. In other words, instead of always looking at the tail of the fish, the user can now look from the side or the front if they choose.